+++ src/main/java/com/slytechs/jnet/protocol/api/address/Ip6Address.java
@@ -155,7 +155,7 @@ public interface Ip6Address extends IpAddress {
 	}
 
 	@Override
-	default int length() {
+	default long length() {
 		return LENGTH;
 	}
 }
