+++ src/main/java/com/slytechs/jnet/protocol/api/address/VlanId.java
@@ -106,7 +106,7 @@ public interface VlanId extends Address {
 	}
 
 	@Override
-	default int length() {
+	default long length() {
 		return LENGTH;
 	}
 }
